Change architecture to io monads after all
For generators like itertools.product, it is possible, that the
function execution and the yielding happen at entirely different
times.
In the special case of itertools.product, the last generator in its
argument list is completely exhausted, once the other iterators are
starting to be consumed.
Therefore, to guarantee that the call stack can be resolved, the
parameters with wich a result was calculated by a function has to
contain said parameters using IO monads.
To ensure that those monads are not visible to the user, each
generator has to be switched to `monadic`, to return a monad instead
of only the result of the function call.
This switching to `monadic` happens when the iterators are created, by
switching all parent iterators to monadic once, `iter` is calld on the
current generator.
This ensures that all parent generators are in the `monadic` state
while the bottom one, which always has the last args and kwargs
anyway, remains in non-monadic mode and returns only the result.